projects/article-summarizer/requirements.txt
anthropic
requests
newspaper3k
pyttsx3
lxml_html_clean
gtts
